backend/pyproject.toml
[tool.poetry]
name = "juliya-nails-fastapi"
version = "0.1.0"
description = ""
authors = ["shadowy-pycoder <shadowy-pycoder@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.11.1"
asyncpg = "^0.28.0"
fastapi = {extras = ["all"], version = "^0.100.0"}
python-dotenv = "^1.0.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.23"}
pydantic-settings = "^2.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
mypy = "^1.7.1"
pydantic = "^2.5.2"
types-passlib = "^1.7.7.13"
types-python-jose = "^3.3.4.8"
passlib = "^1.7.4"
fastapi-pagination = "^0.12.14"
fastapi-filter = {extras = ["sqlalchemy"], version = "^1.0.0"}
pillow = "^10.1.0"
types-pillow = "^10.1.0.2"
aiofiles = "^23.2.1"
types-aiofiles = "^23.2.0.0"
filetype = "^1.2.0"
fastapi-mail = "^1.4.1"
itsdangerous = "^2.1.2"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"
redis = "^5.0.1"
fastapi-cache2 = "^0.2.1"
types-redis = "^4.6.0.20240106"
pre-commit = "3.7.0"


[tool.poetry.group.dev.dependencies]
fastapi2postman = "^0.1.2"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-dotenv = "^0.5.2"
fakeredis = "^2.20.1"
pytest-mock = "^3.12.0"
pytest-dependency = "^0.6.0"
pytest-freezegun = "^0.4.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
cache_dir = ".mypy_cache/strict"
allow_redefinition = true
strict_optional = true
show_error_codes = true
show_column_numbers = true
warn_no_return = true
disallow_any_unimported = true
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
implicit_reexport = true
strict_equality = true
ignore_missing_imports = true


[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true


[tool.ruff]
exclude = [
    ".git",
    ".git-rewrite",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"env.py" = ["E402", "F401"]